third_party/blink/web_tests/external/wpt/css/css-flexbox/parsing/flex-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['flex'] = "calc(10 + (sign(20cqw - 10px) * 5)) calc(10 + (sign(20cqw - 10px) * 5)) 1px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['flex'] = "1 1 calc(10px + (sign(20cqw - 10px) * 5px))" should set the property value
  assert_equals: serialization should be canonical expected "1 1 calc(10px + (5px * sign(20cqw - 10px)))" but got "1 1 calc(10px + sign(20cqw - 10px) * 5px)"
[FAIL] e.style['flex'] = "calc(1) calc(2 + 1) calc(3px)" should set the property value
  assert_equals: serialization should be canonical expected "calc(1) calc(3) calc(3px)" but got "1 3 calc(3px)"
[FAIL] e.style['flex'] = "calc(-1) calc(-1) 0" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
